dto 위치 서비스->dto단으로 변경 및 테스트 코드 일부작성
