+++ src/main/java/shop/mtcoding/finalproject/dto/menu/MenuRespDto.java
@@ -1,11 +1,45 @@
 package shop.mtcoding.finalproject.dto.menu;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import lombok.Getter;
 import lombok.Setter;
 import shop.mtcoding.finalproject.config.enums.MenuCategoryEnum;
 import shop.mtcoding.finalproject.domain.menu.Menu;
 
 public class MenuRespDto {
+
+    /* 성진 작업 시작 */
+    @Getter
+    @Setter
+    public static class MenuListRespDto {
+        private List<MenuDto> menus = new ArrayList<>();
+
+        public MenuListRespDto(List<Menu> menus) {
+            this.menus = menus.stream().map((menu) -> new MenuDto(menu)).collect(Collectors.toList());
+        }
+
+        @Getter
+        @Setter
+        public class MenuDto {
+            private String name;
+            private String intro;
+            private String price;
+            private String thumbnail;
+            private MenuCategoryEnum category;
+
+            public MenuDto(Menu menu) {
+                this.name = menu.getName();
+                this.intro = menu.getIntro();
+                this.price = menu.getPrice();
+                this.thumbnail = menu.getThumbnail();
+                this.category = menu.getCategory();
+            }
+
+        }
+    }
 
     /* 승현 작업 시작 */
